add Exception handling code / organize Dubeg log

+++ src/daemon/mtp_daemon_gdbus_manager.c
@@ -41,7 +41,7 @@ static void __manager_get_device_list_thread_func(gpointer user_data)
 	g_assert(param->invocation != NULL);
 	g_assert(param->mtp_ctx != NULL);
 
-	MTP_LOGI("%s", __func__);
+	/*MTP_LOGI("%s", __func__);*/
 
 	/* parameter unpacking */
 	object = (mtpgdbuslibManager *)(param->object);
@@ -50,7 +50,7 @@ static void __manager_get_device_list_thread_func(gpointer user_data)
 	g_variant_builder_init(&b, G_VARIANT_TYPE("aa{sv}"));
 
 	count = param->mtp_ctx->device_list->device_num;
-	MTP_LOGE("devices_num : %d", count);
+	MTP_LOGI("devices_num: %d", count);
 
 	for (slot = 1; slot < MTP_MAX_SLOT; slot++) {
 		mtp_device_info *device_info;
@@ -94,7 +94,7 @@ static void __manager_get_device_handle_thread_func(gpointer user_data)
 	g_assert(param->object != NULL);
 	g_assert(param->invocation != NULL);
 
-	MTP_LOGI("%s", __func__);
+	/*MTP_LOGI("%s", __func__);*/
 
 	/* parameter unpacking */
 	busno = (int)(param->param1);
@@ -105,17 +105,19 @@ static void __manager_get_device_handle_thread_func(gpointer user_data)
 		mtp_device_info *device_info = NULL;
 		device_info = param->mtp_ctx->device_list->device_info_list[slot];
 
-		if (busno == device_info->bus_location) {
+		if (device_info != NULL && busno == device_info->bus_location) {
 			handle = slot;
-			MTP_LOGI("handle : %d, device : %p", handle, device_info->device);
+			MTP_LOGI("handle: %d, device: %p", handle, device_info->device);
 			break;
 		}
 	}
 
 	if (handle != -1)
-		MTP_LOGE("found MTP device - handle : %p, bus: %d", handle, busno);
-	else
+		MTP_LOGI("found MTP device - handle: %p, bus: %d", handle, busno);
+	else {
 		MTP_LOGE("!!! not found MTP device handle - bus: %d", busno);
+		result = MTP_ERROR_NO_DEVICE;
+	}
 
 	mtp_gdbuslib_manager_complete_get_device_handle(object, param->invocation, handle, result);
 
@@ -130,8 +132,10 @@ static void __manager_get_storage_ids_thread_func(gpointer user_data)
 	/* variable definition */
 	mtp_param *param = (mtp_param *)user_data;
 	mtp_error_e result = MTP_ERROR_NONE;
-	LIBMTP_mtpdevice_t *device;
+	LIBMTP_mtpdevice_t *device = NULL;
 	LIBMTP_devicestorage_t *storage;
+	mtp_device_info *device_info;
+	int device_id;
 	int count = 0;
 
 	GVariant *gv = NULL;
@@ -143,31 +147,41 @@ static void __manager_get_storage_ids_thread_func(gpointer user_data)
 	g_assert(param->invocation != NULL);
 	g_assert(param->mtp_ctx != NULL);
 
-	MTP_LOGI("%s", __func__);
+	/*MTP_LOGI("%s", __func__);*/
 
 	/* parameter unpacking */
-	device = (LIBMTP_mtpdevice_t *)param->mtp_ctx->device_list->device_info_list[param->param1]->device;
-	MTP_LOGI("param->param1 %d, device %p", param->param1, device);
+	device_id = param->param1;
+	device_info = (mtp_device_info *)param->mtp_ctx->device_list->device_info_list[device_id];
+
+	if (device_info != NULL)
+		device = (LIBMTP_mtpdevice_t *)device_info->device;
+
+	MTP_LOGI("device_id: %d, device: %p", device_id, device);
 
 	/* loop over storages */
 	g_variant_builder_init(&b, G_VARIANT_TYPE("aa{sv}"));
 
-	for (storage = device->storage; storage != NULL; storage = storage->next) {
-		MTP_LOGE("storage : %p", storage);
+	if (device_info != NULL) {
+		for (storage = device->storage; storage != NULL; storage = storage->next) {
+			MTP_LOGI("storage: %p", storage);
 
-		g_variant_builder_open(&b, G_VARIANT_TYPE("a{sv}"));
+			g_variant_builder_open(&b, G_VARIANT_TYPE("a{sv}"));
 
-		g_variant_builder_add(&b, "{sv}", "storage_id",
-			g_variant_new_int32(storage->id));
+			g_variant_builder_add(&b, "{sv}", "storage_id",
+				g_variant_new_int32(storage->id));
 
-		g_variant_builder_close(&b);
+			g_variant_builder_close(&b);
 
-		count++;
+			count++;
+		}
+	} else {
+		MTP_LOGE("!!! no MTP device");
+		result = MTP_ERROR_NO_DEVICE;
 	}
 
 	gv = g_variant_builder_end(&b);
 
-	MTP_LOGE("storage_num : %d", count);
+	MTP_LOGI("storage_num: %d", count);
 
 	mtp_gdbuslib_manager_complete_get_storage_ids(param->object,
 		param->invocation, count, gv, result);
@@ -183,10 +197,12 @@ static void __manager_get_object_handles_thread_func(gpointer user_data)
 	/* variable definition */
 	mtp_param *param = (mtp_param *)user_data;
 	mtp_error_e result = MTP_ERROR_NONE;
+	mtp_device_info *device_info;
+	LIBMTP_mtpdevice_t *device = NULL;
 
-	LIBMTP_mtpdevice_t *device;
 	int ret;
 	int slot;
+	int device_id;
 	int storage_id;
 	int format;
 	int parent_object_handle;
@@ -206,76 +222,88 @@ static void __manager_get_object_handles_thread_func(gpointer user_data)
 	/*MTP_LOGI("%s", __func__);*/
 
 	/* parameter unpacking */
-	device = (LIBMTP_mtpdevice_t *)param->mtp_ctx->device_list->device_info_list[param->param1]->device;
-	MTP_LOGI("param->param1 %d, device %p", param->param1, device);
-
+	device_id = param->param1;
 	storage_id = param->param2;
 	format = param->param3;
 	parent_object_handle = param->param4;
 
-	MTP_LOGI("storage id : %d", param->param2);
-	MTP_LOGI("format : %d", param->param3);
-	MTP_LOGI("parent : %d", param->param4);
+	device_info = (mtp_device_info *)param->mtp_ctx->device_list->device_info_list[device_id];
 
-	ret = LIBMTP_Get_Object_Handles(device, (uint32_t)storage_id, (uint32_t)format, (uint32_t)parent_object_handle,
-		(uint32_t **)&object_list, (uint32_t *)&total_object_num);
+	if (device_info != NULL)
+		device = (LIBMTP_mtpdevice_t *)device_info->device;
 
-	if (ret < 0)
-		MTP_LOGE("LIBMTP_Get_Object_Handles is failed : %d", ret);
+	MTP_LOGI("device_id: %d, storage id: %d, format: %d, parent: %d, device: %p",
+		param->param1, param->param2, param->param3, param->param4, device);
 
-	/*MTP_LOGI("Total Object number : %d", temp);*/
+	if (device_info != NULL) {
+		ret = LIBMTP_Get_Object_Handles(device, (uint32_t)storage_id,
+			(uint32_t)format, (uint32_t)parent_object_handle,
+			(uint32_t **)&object_list, (uint32_t *)&total_object_num);
+
+		if (ret < 0)
+			MTP_LOGE("LIBMTP_Get_Object_Handles is failed : %d", ret);
+	}
+
+	/*MTP_LOGI("Total Object number: %d", temp);*/
 
 	/* loop over storages */
 	g_variant_builder_init(&b, G_VARIANT_TYPE("aa{sv}"));
 
-	mtp_daemon_db_begin(param->mtp_ctx);
+	if (device_info != NULL) {
+		mtp_daemon_db_begin(param->mtp_ctx);
 
-	for (slot = 0; slot < total_object_num; slot++) {
-		MTPObjectInfo *object_info = NULL;
-		object_info = LIBMTP_Get_Object_Info(device, (uint32_t)object_list[slot]);
-
-		if (object_info != NULL)
-			mtp_daemon_db_insert(param->param1, storage_id, object_list[slot], object_info, param->mtp_ctx);
-
-		/*if (mtp_daemon_db_is_exist((int)device, object_list[i], param->mtp_ctx) == false) {
-			MTP_LOGI("Objectinfo is not stored. %d, %d", device, object_list[i]);
-			object_info = LIBMTP_Get_Object_Info(device, object_list[i]);
+		for (slot = 0; slot < total_object_num; slot++) {
+			MTPObjectInfo *object_info = NULL;
+			/*object_info = LIBMTP_Get_Object_Info(device, (uint32_t)object_list[slot]);
 
 			if (object_info != NULL)
-				mtp_daemon_db_insert(device, storage_id, object_list[i], object_info, param->mtp_ctx);
-		} else {
-			MTP_LOGI("Objectinfo is stored. ");
-			object_info = mtp_daemon_db_get_object_info(device, object_list[i], param->mtp_ctx);
-		}*/
+				mtp_daemon_db_insert(device_id, storage_id, object_list[slot], object_info, param->mtp_ctx);*/
 
-		if (object_info != NULL) {
-			/*MTP_LOGI("object id : %d, format : %d, filename : %s", object_list[i],
-							object_info->ObjectFormat, object_info->Filename);*/
+			if (mtp_daemon_db_is_exist(device_id, object_list[slot], param->mtp_ctx) == false) {
+				MTP_LOGI("Objectinfo is Not stored. device_id: %d, object_list[%d]: %d",
+					device_id, slot, object_list[slot]);
+				object_info = LIBMTP_Get_Object_Info(device, (uint32_t)object_list[slot]);
 
-			if ((format == LIBMTP_FILETYPE_ALL &&
-					(object_info->ObjectFormat == LIBMTP_FILETYPE_FOLDER ||
-					LIBMTP_FILETYPE_IS_IMAGE(object_info->ObjectFormat))) ||
-				(format == LIBMTP_FILETYPE_ALL_IMAGE &&
-					LIBMTP_FILETYPE_IS_IMAGE(object_info->ObjectFormat)) ||
-				(format == object_info->ObjectFormat)) {
-				g_variant_builder_open(&b, G_VARIANT_TYPE("a{sv}"));
-
-				g_variant_builder_add(&b, "{sv}", "object_handle",
-				g_variant_new_int32(object_list[slot]));
-
-				g_variant_builder_close(&b);
-
-				target_object_num++;
+				if (object_info != NULL)
+					mtp_daemon_db_insert(device_id, storage_id, object_list[slot], object_info, param->mtp_ctx);
+			} else {
+				MTP_LOGI("Objectinfo is stored. device_id: %d, object_list[%d]: %d",
+					device_id, slot, object_list[slot]);
+				object_info = mtp_daemon_db_get_object_info(device_id, object_list[slot], param->mtp_ctx);
 			}
-		}
-		g_free(object_info);
-	}
 
-	mtp_daemon_db_commit(param->mtp_ctx);
+			if (object_info != NULL) {
+				/*MTP_LOGI("object id: %d, format : %d, filename : %s", object_list[slot],
+					object_info->ObjectFormat, object_info->Filename);*/
+
+				if ((format == LIBMTP_FILETYPE_ALL &&
+						(object_info->ObjectFormat == LIBMTP_FILETYPE_FOLDER ||
+						LIBMTP_FILETYPE_IS_IMAGE(object_info->ObjectFormat))) ||
+					(format == LIBMTP_FILETYPE_ALL_IMAGE &&
+						LIBMTP_FILETYPE_IS_IMAGE(object_info->ObjectFormat)) ||
+					(format == object_info->ObjectFormat)) {
+					g_variant_builder_open(&b, G_VARIANT_TYPE("a{sv}"));
+
+					g_variant_builder_add(&b, "{sv}", "object_handle",
+					g_variant_new_int32(object_list[slot]));
+
+					g_variant_builder_close(&b);
+
+					target_object_num++;
+				}
+			}
+			g_free(object_info);
+		}
+
+		mtp_daemon_db_commit(param->mtp_ctx);
+	} else {
+		MTP_LOGE("!!! no MTP device");
+		result = MTP_ERROR_NO_DEVICE;
+	}
 
 	gv = g_variant_builder_end(&b);
 
-	/*MTP_LOGE("Real object_num : %d", target_object_num);*/
+	/*MTP_LOGI("Real object_num: %d", target_object_num);*/
 
 	mtp_gdbuslib_manager_complete_get_object_handles(param->object,
 		param->invocation, target_object_num, gv, result);
@@ -291,12 +319,14 @@ static void __manager_get_object_handles_thread_func(gpointer user_data)
 static void __manager_get_object_thread_func(gpointer user_data)
 {
 	/* variable definition */
-	int ret;
-	int fd;
 	mtp_param *param = (mtp_param *)user_data;
 	mtp_error_e result = MTP_ERROR_NONE;
+	int ret;
+	int fd;
 
 	LIBMTP_mtpdevice_t *device;
+	mtp_device_info *device_info;
+	int device_id;
 	int object_handle;
 	char *dest_path;
 
@@ -309,34 +339,42 @@ static void __manager_get_object_thread_func(gpointer user_data)
 	/*MTP_LOGI("%s", __func__);*/
 
 	/* parameter unpacking */
-	device = (LIBMTP_mtpdevice_t *)param->mtp_ctx->device_list->device_info_list[param->param1]->device;
-	MTP_LOGI("param->param1 %d, device %p", param->param1, device);
-
+	device_id = param->param1;
 	object_handle = param->param2;
 	dest_path = param->char_param1;
 
-	/* file open */
-	if (access(dest_path, F_OK) >= 0) {
-		MTP_LOGE("file exist - dest_path : %s", dest_path);
+	device_info = (mtp_device_info *)param->mtp_ctx->device_list->device_info_list[device_id];
 
-		if (unlink(dest_path) < 0)
-			MTP_LOGE("remove fail - dest_path : %s", dest_path);
-	}
+	if (device_info != NULL) {
+		device = (LIBMTP_mtpdevice_t *)device_info->device;
+		MTP_LOGI("device_id: %d, device: %p, dest_path: %s", device_id, device, dest_path);
 
-	fd = open(dest_path, O_WRONLY | O_CREAT | O_EXCL, 0644);
+		/* file open */
+		if (access(dest_path, F_OK) >= 0) {
+			MTP_LOGE("file exist - dest_path: %s", dest_path);
 
-	if (fd < 0) {
-		MTP_LOGE("%s file open fail - fd : %d, dest_path : %s", __func__, fd, dest_path);
-		result = MTP_ERROR_GENERAL;
-	} else {
-		ret = LIBMTP_Get_File_To_File_Descriptor(device, object_handle, fd, NULL, NULL);
-
-		if (ret != 0) {
-			MTP_LOGE("get descriptor fail - ret : %d", ret);
-			result = MTP_ERROR_PLUGIN;
+			if (unlink(dest_path) < 0)
+				MTP_LOGE("remove fail - dest_path: %s", dest_path);
 		}
 
-		close(fd);
+		fd = open(dest_path, O_WRONLY | O_CREAT | O_EXCL, 0644);
+
+		if (fd < 0) {
+			MTP_LOGE("file open fail - fd: %d, dest_path: %s", fd, dest_path);
+			result = MTP_ERROR_GENERAL;
+		} else {
+			ret = LIBMTP_Get_File_To_File_Descriptor(device, object_handle, fd, NULL, NULL);
+
+			if (ret != 0) {
+				MTP_LOGE("get descriptor fail - ret: %d", ret);
+				result = MTP_ERROR_PLUGIN;
+			}
+
+			close(fd);
+		}
+	} else {
+		MTP_LOGE("!!! no MTP device");
+		result = MTP_ERROR_NO_DEVICE;
 	}
 
 	mtp_gdbuslib_manager_complete_get_object(param->object,
@@ -353,14 +391,16 @@ static void __manager_get_object_thread_func(gpointer user_data)
 static void __manager_get_thumbnail_thread_func(gpointer user_data)
 {
 	/* variable definition */
-	int ret;
-	int fd;
 	mtp_param *param = (mtp_param *)user_data;
 	mtp_error_e result = MTP_ERROR_NONE;
 	unsigned char *thumb_data;
 	unsigned int thumb_size;
+	int ret;
+	int fd;
 
 	LIBMTP_mtpdevice_t *device;
+	mtp_device_info *device_info;
+	int device_id;
 	int object_handle;
 	char *dest_path;
 
@@ -373,40 +413,46 @@ static void __manager_get_thumbnail_thread_func(gpointer user_data)
 	/*MTP_LOGI("%s", __func__);*/
 
 	/* parameter unpacking */
-	device = (LIBMTP_mtpdevice_t *)param->mtp_ctx->device_list->device_info_list[param->param1]->device;
-	MTP_LOGI("param->param1 %d, device %p", param->param1, device);
-
+	device_id = param->param1;
 	object_handle = param->param2;
 	dest_path = param->char_param1;
 
-	/*MTP_LOGI("dest_path : %s", dest_path);*/
+	device_info = (mtp_device_info *)param->mtp_ctx->device_list->device_info_list[device_id];
 
-	/* file open */
-	if (access(dest_path, F_OK) >= 0) {
-		MTP_LOGE("file exist - dest_path : %s", dest_path);
+	if (device_info != NULL) {
+		device = (LIBMTP_mtpdevice_t *)device_info->device;
+		MTP_LOGI("device_id: %d, device: %p, dest_path: %s", device_id, device, dest_path);
 
-		if (unlink(dest_path) < 0)
-			MTP_LOGE("remove fail - dest_path : %s", dest_path);
-	}
+		/* file open */
+		if (access(dest_path, F_OK) >= 0) {
+			MTP_LOGE("file exist - dest_path: %s", dest_path);
 
-	fd = open(dest_path, O_WRONLY | O_CREAT | O_EXCL, 0644);
-
-	if (fd < 0) {
-		MTP_LOGE("%s file open fail - fd : %d, dest_path : %s", __func__, fd, dest_path);
-		result = MTP_ERROR_GENERAL;
-	} else {
-		ret = LIBMTP_Get_Thumbnail(device, object_handle, &thumb_data, &thumb_size);
-
-		if (ret == 0) {
-			ret = write(fd, thumb_data, thumb_size);
-			if (ret == 0)
-				result = MTP_ERROR_ALLOC_FAIL;
-		} else {
-			MTP_LOGE("get thumbnail fail - ret : %d", ret);
-			result = MTP_ERROR_PLUGIN;
+			if (unlink(dest_path) < 0)
+				MTP_LOGE("remove fail - dest_path: %s", dest_path);
 		}
 
-		close(fd);
+		fd = open(dest_path, O_WRONLY | O_CREAT | O_EXCL, 0644);
+
+		if (fd < 0) {
+			MTP_LOGE("file open fail - fd: %d, dest_path: %s", fd, dest_path);
+			result = MTP_ERROR_GENERAL;
+		} else {
+			ret = LIBMTP_Get_Thumbnail(device, object_handle, &thumb_data, &thumb_size);
+
+			if (ret == 0) {
+				ret = write(fd, thumb_data, thumb_size);
+				if (ret == 0)
+					result = MTP_ERROR_ALLOC_FAIL;
+			} else {
+				MTP_LOGE("get thumbnail fail - ret: %d", ret);
+				result = MTP_ERROR_PLUGIN;
+			}
+
+			close(fd);
+		}
+	} else {
+		MTP_LOGE("!!! no MTP device");
+		result = MTP_ERROR_NO_DEVICE;
 	}
 
 	mtp_gdbuslib_manager_complete_get_thumbnail(param->object,
@@ -428,6 +474,8 @@ static void __manager_delete_object_thread_func(gpointer user_data)
 	mtp_error_e result = MTP_ERROR_NONE;
 
 	LIBMTP_mtpdevice_t *device;
+	mtp_device_info *device_info;
+	int device_id;
 	int object_handle;
 
 	/* check precondition */
@@ -439,14 +487,23 @@ static void __manager_delete_object_thread_func(gpointer user_data)
 	MTP_LOGI("%s", __func__);
 
 	/* parameter unpacking */
-	device = (LIBMTP_mtpdevice_t *)param->mtp_ctx->device_list->device_info_list[param->param1]->device;
-	MTP_LOGI("param->param1 %d, device %p", param->param1, device);
+	device_id = param->param1;
+	device_info = (mtp_device_info *)param->mtp_ctx->device_list->device_info_list[device_id];
 
-	object_handle = param->param2;
+	if (device_info != NULL) {
+		device = (LIBMTP_mtpdevice_t *)device_info->device;
 
-	ret = LIBMTP_Delete_Object(device, object_handle);
-	if (ret != 0)
-		result = MTP_ERROR_PLUGIN;
+		MTP_LOGI("device_id: %d, device: %p", device_id, device);
+
+		object_handle = param->param2;
+
+		ret = LIBMTP_Delete_Object(device, object_handle);
+		if (ret != 0)
+			result = MTP_ERROR_PLUGIN;
+	} else {
+		MTP_LOGE("!!! no MTP device");
+		result = MTP_ERROR_NO_DEVICE;
+	}
 
 	mtp_gdbuslib_manager_complete_delete_object(param->object,
 		param->invocation, result);
@@ -466,7 +523,7 @@ gboolean manager_get_device_list(
 	mtp_param *param = NULL;
 	gint result = MTP_ERROR_NONE;
 
-	MTP_LOGE(">>> REQUEST from [%s]",
+	MTP_LOGI(">>> REQUEST from [%s]",
 		g_dbus_method_invocation_get_sender(invocation));
 
 	param = g_try_new0(mtp_param, 1);
@@ -513,7 +570,7 @@ gboolean manager_get_device_handle(
 	mtp_param *param = NULL;
 	gint result = MTP_ERROR_NONE;
 
-	MTP_LOGE(">>> REQUEST from [%s]",
+	MTP_LOGI(">>> REQUEST from [%s]",
 		g_dbus_method_invocation_get_sender(invocation));
 
 	param = g_try_new0(mtp_param, 1);
@@ -561,7 +618,7 @@ gboolean manager_get_storage_ids(
 	mtp_param *param = NULL;
 	gint result = MTP_ERROR_NONE;
 
-	MTP_LOGE(">>> REQUEST from [%s]",
+	MTP_LOGI(">>> REQUEST from [%s]",
 		g_dbus_method_invocation_get_sender(invocation));
 
 	param = g_try_new0(mtp_param, 1);
@@ -773,7 +830,7 @@ gboolean manager_delete_object(
 	mtp_param *param = NULL;
 	gint result = MTP_ERROR_NONE;
 
-	MTP_LOGE(">>> REQUEST from [%s]",
+	MTP_LOGI(">>> REQUEST from [%s]",
 		g_dbus_method_invocation_get_sender(invocation));
 
 	param = g_try_new0(mtp_param, 1);
